+++ doc/changelog.rst
@@ -27,6 +27,7 @@ Localization
 Admin
 ~~~~~
 
+- Manage category products from category edit view
 - Filter products based on category
 - Add permission check for dashboard blocks
 - Fix required permission issues for various modules